The beginnings of a badge
This is a rough "skeleton" of what the OSSEM-Badge should be. Currently the LCD is the main concern. It loads a bitmap of the OSSEM logo and cycles through the screen pixel values. 

A lot of the code is unneeded. This is because we have used the DC-Darknet badge as an example.